Update: penambahan halaman perhitungan, batch, dan pemisah pengajuan

+++ project/database/migrations/2025_09_30_112140_create_submissions_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->float('revenue');
             $table->float('asset');
             $table->integer('dependents');
-            $table->integer('building_area');
+            $table->float('building_area');
             $table->integer('building_legality');
             $table->integer('roof_condition');
             $table->integer('wall_condition');
